pyproject.toml
[project]
name = "celery-decipher"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "celery[pytest]>=5.5.0",
    "fastapi>=0.115.12",
    "langdetect>=1.0.9",
    "opentelemetry-api>=1.32.0",
    "opentelemetry-exporter-otlp>=1.32.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.32.0",
    "opentelemetry-exporter-otlp-proto-http>=1.32.0",
    "opentelemetry-instrumentation-celery>=0.53b0",
    "opentelemetry-instrumentation-fastapi>=0.53b0",
    "opentelemetry-instrumentation-logging>=0.53b0",
    "opentelemetry-instrumentation-psycopg>=0.53b0",
    "opentelemetry-instrumentation-starlette>=0.53b0",
    "opentelemetry-sdk>=1.32.0",
    "psycopg[binary,pool]>=3.2.6",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
    "rapidfuzz>=3.13.0",
    "sqlalchemy>=2.0.40",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.28.5",
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.11.4",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
